requirements.txt
streamlit
pandas
streamlit-oauth
pyjwt
python-dotenv